Remove Nashorn and re-use HttpClient
Nashorn is no longer supported, so get rid of it in the codebase.

Remove stuff that was used to build the javascript version (e.g.
package.json, node plugin in pom.xml).

Re-use the HttpClient – it's not meant to be created for every request,
we don't get the advantage of connection pooling and keepalive that way.
Doing this required converting all the static method calls to normal
object initialisation, but keeps the (static) client interface in
DosisTilTekstWrapper intact.

Get rid of legacy junk (trifork-common, Eclipse settings).